rename and moving displays for folder
